Part C and XFintoARA.py
1) Removed XFintoARA2.py (only differs from XFintoARA.py by one line)
2) Setting up Part C skeleton in main.sh
3) TODO: rewrite XFintoARA.py -> part_c.py
