--- Evolutionary_loop/Antenna_Performance_Metric/name.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-      
-for file in *.sh;
-do
-  name=`echo $file | sed "s+${file}+\.${file}+"`
-  mv $file $name
-done
